fix: data-safety blockers (recovery scan, CRC check, u64 generation, partial reads)
Seven correctness fixes to the storage engine:

1. DataFile::open no longer trusts the (never-updated) persisted
   wblock_count header. Scans physical blocks on open so flushed data
   files survive a restart even if the WAL has been trimmed.
2. DirectFile write_at / read_at now loop over pwrite / pread until the
   full buffer has been transferred and retry EINTR — single-syscall
   short writes no longer silently truncate WBlocks.
3. WAL replay verifies each entry's CRC32 and stops at the first
   corrupt / torn entry. Torn tails can no longer produce bogus
   index updates that survive future durable writes.
4. Recovery rebuilds WBlockMeta (live / total records, bytes_used) as
   it scans each block. Compaction's get_fragmented_blocks now returns
   real data after a restart instead of always seeing an empty table.
5. DiskLocation carries record_size through the write path, and the
   Handler GET path uses partial pread for known-size records —
   small cold GETs no longer always fetch a full 1 MiB WBlock.
6. Generation counter widened to u64 end-to-end (record header, WAL
   entry header, index, WAL state, recovery stats). At 10M writes/sec
   it now takes 58k years to wrap instead of ~71 minutes.
   On-disk formats bumped accordingly: record header 32 → 40 bytes,
   WAL entry header 24 → 28 bytes, WAL magic 'WAL!' → 'WAL2'.
7. put_nowait_on documents the index-before-durable visibility window
   so callers know to use put_sync when they need strict
   visible ⇒ durable.

Tests: all 278 unit + integration tests green. Added a regression
test that DataFile::open recovers the correct wblock count even when
the header is stale.

+++ benches/throughput.rs
@@ -104,7 +104,7 @@ fn bench_index_operations(c: &mut Criterion) {
     for i in 0..100000 {
         let key = format!("index_key_{:08}", i);
         let location = ssd_kv::storage::write_buffer::DiskLocation::new(0, 0, i as u32);
-        index.insert(key.as_bytes(), location, i as u32, 100);
+        index.insert(key.as_bytes(), location, i as u64, 100);
     }
 
     let mut group = c.benchmark_group("index");
@@ -125,7 +125,7 @@ fn bench_index_operations(c: &mut Criterion) {
             i = i.wrapping_add(1);
             let key = format!("new_key_{}", i);
             let location = ssd_kv::storage::write_buffer::DiskLocation::new(0, 0, i);
-            index.insert(key.as_bytes(), location, i, 100)
+            index.insert(key.as_bytes(), location, i as u64, 100)
         })
     });
